Move `@tina/page` to server components
Don't need to render this client-side
